pyproject.toml
[project]
name = "sign-language-recognition-api"
version = "0.1.0"
description = "FastAPI backend for real-time sign language recognition using CNN (MobileNet) with live video inference."
authors = [
    { name = "Mithun Thomas", email = "mithun@example.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11, <4"

# ---------------- CORE BACKEND DEPENDENCIES ----------------
dependencies = [
    # FastAPI stack
    "fastapi>=0.109.1",
    "uvicorn>=0.27.0",
    "uvloop>=0.19.0",
    "httptools>=0.6.1",

    # Config & validation
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.6.1",
    "pydantic-settings>=2.0.3",

    # Auth & security
    "python-jose>=3.3.0",
    "bcrypt>=4.1.1",
    "python-multipart>=0.0.9",

    # Database
    "SQLAlchemy>=2.0.25",
    "SQLAlchemy-Utils>=0.41.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "greenlet>=2.0.2",

    # CRUD & Admin
    "fastcrud>=0.19.2",
    "crudadmin>=0.4.2",

    # Background jobs & cache
    "redis>=5.0.1",
    "arq>=0.25.0",

    # HTTP client
    "httpx>=0.26.0",

    # Runtime
    "gunicorn>=23.0.0",
# 🔬 Machine Learning (CPU-safe)

    "tensorflow-cpu==2.15.0",
    "numpy==1.26.4",
    "opencv-python-headless==4.9.0.80",
    "scikit-learn>=1.4.0",
    "pillow>=10.2.0",
]

# ---------------- OPTIONAL DEPENDENCIES ----------------
[project.optional-dependencies]



# 🧪 Development & Testing
dev = [
    "pytest>=7.4.2",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=1.0.0",
    "faker>=26.0.0",
    "pre-commit>=4.3.0",

    # Type checking & linting
    "mypy>=1.16.0",
    "types-redis>=4.6.0",
    "ruff>=0.11.13",
]

# ---------------- BUILD SYSTEM ----------------
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["src/"]

[tool.hatch.build.targets.wheel]
include = ["src/"]
packages = ["src"]

# ---------------- RUFF (LINTING) ----------------
[tool.ruff]
target-version = "py311"
line-length = 120
fix = true

[tool.ruff.lint]
select = ["F", "E", "W", "C", "UP", "I"]
ignore = [
    "E402",
    "UP006",
    "UP007",
    "E741",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.mccabe]
max-complexity = 24

[tool.ruff.lint.pydocstyle]
convention = "numpy"

# ---------------- PYTEST ----------------
[tool.pytest.ini_options]
filterwarnings = [
    "ignore::PendingDeprecationWarning:starlette.formparsers",
]

# ---------------- MYPY ----------------
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
mypy_path = "src"
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "src.app.*"
disallow_untyped_defs = true
